+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **add:** documentation blocks for helper functions
 - **upg:** eslint (2.9.0 -> 2.10.1)
 - **fix:** lint
+- **fix:** rename internals.schema into internals.schemata
 
 ### [v0.2.0] - 2016-05-15
 - **fix:** change eslint (no-param-reassign) rule